server_django/requirements.txt
django>=4.2
django-cors-headers>=4.3
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
daphne>=4.0
